--- src/services/cluster/peers/address.rs
@@ -1,14 +1,0 @@
-use crate::services::cluster::peers::identifier::PeerIdentifier;
-use crate::{from_to, make_smart_pointer};
-
-make_smart_pointer!(PeerAddrs, Vec<PeerIdentifier>);
-from_to!(Vec<PeerIdentifier>, PeerAddrs);
-
-#[derive(Debug)]
-pub struct PeerAddrs(pub Vec<PeerIdentifier>);
-
-impl PeerAddrs {
-    pub fn stringify(self) -> String {
-        self.0.into_iter().map(|x| x.0).collect::<Vec<String>>().join(" ")
-    }
-}
