+++ src/services/cluster/peers/mod.rs
@@ -1,5 +1,3 @@
-pub(crate) mod address;
-
 pub(crate) mod connected_types;
 pub(crate) mod identifier;
 pub(crate) mod kind;
